refactor: Migrate weather API handler to align with provider-based architecture
